requirements.txt
defusedxml==0.6.0
diff-match-patch==20181111
Django==2.2.1
django-cors-headers==3.2.1
django-import-export==2.0.2
djangorestframework==3.11.0
et-xmlfile==1.0.1
jdcal==1.4.1
MarkupPy==1.14
mysqlclient==1.4.6
odfpy==1.4.1
openpyxl==3.0.3
Pillow==7.0.0
pytz==2019.1
PyYAML==5.3.1
sqlparse==0.3.0
tablib==1.1.0
xlrd==1.2.0
xlwt==1.3.0
